+++ pik/core/models/_collector_delete.py
@@ -1,4 +1,3 @@
-# pylint: skip-file
 from collections import Counter
 from django.db.models.expressions import Col
 from django.utils.timezone import now
@@ -9,12 +8,28 @@ from django.db.models import signals, sql
 from django.db.models.deletion import Collector
 from django.db.models.fields.related import ForeignObject
 from django.utils import six
+from django.utils.translation import ugettext_lazy as _
+from django.conf import settings
 
 FIELD = 'deleted'
 
 
+class DeleteNotSoftDeletedModel(Exception):
+    pass
+
+
+def _is_soft_excluded(model):
+    soft_delete_exclude_list = getattr(settings, 'SOFT_DELETE_EXCLUDE', [])
+    value = f'{model._meta.app_label}.{model._meta.object_name}'
+    if value in soft_delete_exclude_list:
+        return True
+    return False
+
+
 def _delete(self):
     from .soft_deleted import SoftDeleted
+    safe_mode = getattr(settings, 'SOFT_DELETE_SAFE_MODE', True)
+
     time = now()
 
     # sort instance collections
@@ -31,7 +46,16 @@ def _delete(self):
     with transaction.atomic(using=self.using, savepoint=False):
         # send pre_delete signals
         for model, obj in self.instances_with_model():
-            if not model._meta.auto_created:
+            if not model._meta.auto_created and not issubclass(  # noqa: pylint=protected-access
+                    model, SoftDeleted):
+
+                if safe_mode and not _is_soft_excluded(model):
+                    raise DeleteNotSoftDeletedModel(
+                        _(f'You are trying to delete {model._meta.object_name} instance,'
+                          f' but {model._meta.object_name} is not subclass of '
+                          f'{SoftDeleted._meta.object_name}. You need to inherit your '
+                          f'model from {SoftDeleted._meta.object_name}'
+                          f' or set settings.SOFT_DELETE_SAFE_MODE to False'))
                 signals.pre_delete.send(
                     sender=model, instance=obj, using=self.using
                 )
@@ -71,13 +95,13 @@ def _delete(self):
                 query = sql.DeleteQuery(model)
                 pk_list = [obj.pk for obj in instances]
                 count = query.delete_batch(pk_list, self.using)
-            deleted_counter[model._meta.label] += count
 
-            if not model._meta.auto_created:
-                for obj in instances:
-                    signals.post_delete.send(
-                        sender=model, instance=obj, using=self.using
-                    )
+                if not model._meta.auto_created:
+                    for obj in instances:
+                        signals.post_delete.send(
+                            sender=model, instance=obj, using=self.using
+                        )
+            deleted_counter[model._meta.label] += count
 
     # update collected instances
     for model, instances_for_fieldvalues in six.iteritems(self.field_updates):