+++ pik/core/models/fields/__init__.py
@@ -1,0 +1,5 @@
+from .inheritance import InheritPrimaryUidField
+
+__all__ = [
+    'InheritPrimaryUidField',
+]