+++ test_core_models/tests/test_models_base.py
@@ -21,8 +21,10 @@ def test_historical_protocol(historical_model):
     obj_first = model.objects.first()
     assert obj_first.pk == obj1.pk
     assert obj_last.pk == obj2.pk
-    assert obj_first.uid < obj_last.uid
-    assert obj_first.pk < obj_last.pk
+    assert obj_first.created < obj_last.created
+    assert obj_first.updated < obj_last.updated
+    assert obj_first.version == obj_last.version
+    assert obj_first.uid != obj_last.uid
 
 
 def test_historical_protocol_fields(historical_model):
@@ -54,7 +56,8 @@ def test_historical_protocol_update(historical_model):
     assert created1 == created2 == created3
 
 
-def test_historical_protocol_history(historical_model):
+def test_historical_protocol_history(historical_model, settings):
+    settings.SOFT_DELETE_SAFE_MODE = False
     model, factory = historical_model
     obj1 = factory.create()
     obj2 = factory.create()